rename _clamp.py to clamp.py
