+++ skfp/fingerprints/neural/__init__.py
@@ -1,4 +1,4 @@
 try:
-    from ._clamp import CLAMPFingerprint
+    from .clamp import CLAMPFingerprint
 except ImportError:
     pass
